Do not copy files or folders again, if a higher hierarchy folder was already copied.

+++ modules/nextflow/src/test/groovy/nextflow/executor/BashWrapperBuilderTest.groovy
@@ -405,7 +405,9 @@ class BashWrapperBuilderTest extends Specification {
                 pathes=`ls -1d test.bam test.bai | sort | uniq`
                 set -f
                 for name in \$pathes; do
-                    sh -c 'mkdir -p "/work/dir/`dirname \\"\$1\\"`"; cp -fRL "\$1" "/work/dir/`dirname \\"\$1\\"`";' _ "\$name" || true
+                    if [[ ! -e "/work/dir/\$name" ]]; then
+                        sh -c 'mkdir -p "/work/dir/`dirname \\"\$1\\"`"; cp -fRL "\$1" "/work/dir/`dirname \\"\$1\\"`";' _ "\$name" || true
+                    fi
                 done
                 set +f
                 shopt -u globstar extglob || true
@@ -427,7 +429,9 @@ class BashWrapperBuilderTest extends Specification {
                 pathes=`ls -1d test.bam test.bai | sort | uniq`
                 set -f
                 for name in \$pathes; do
-                    sh -c 'mkdir -p "/another/dir/`dirname \\"\$1\\"`"; mv "\$1" "/another/dir/`dirname \\"\$1\\"`";' _ "\$name" || true
+                    if [[ ! -e "/another/dir/\$name" ]]; then
+                        sh -c 'mkdir -p "/another/dir/`dirname \\"\$1\\"`"; mv "\$1" "/another/dir/`dirname \\"\$1\\"`";' _ "\$name" || true
+                    fi
                 done
                 set +f
                 shopt -u globstar extglob || true
